19 may changes1
19 may changes1

+++ products/views.py
@@ -1,8 +1,8 @@
 
 from flask import render_template, redirect, url_for, flash, request
-from flask_login import login_required
+from flask_login import login_required, current_user
 from .routes import products_bp
-from models import db, Product
+from models import db, Product, ActivityLog
 
 @products_bp.route('/', methods=['GET', 'POST'])
 @login_required
@@ -13,9 +13,14 @@ def crud_products():
 		# Delete
 		delete_id = request.form.get('delete_id', type=int)
 		if delete_id:
+			if current_user.role != 'admin':
+				flash('Only admin can delete products.', 'danger')
+				return redirect(url_for('products.crud_products'))
 			product = Product.query.get(delete_id)
 			if product:
 				db.session.delete(product)
+				db.session.commit()
+				db.session.add(ActivityLog(user_id=current_user.id, username=current_user.username, action='Delete Product', details=f'Product: {product.name}'))
 				db.session.commit()
 				flash('Product deleted!', 'success')
 			return redirect(url_for('products.crud_products'))
@@ -33,15 +38,26 @@ def crud_products():
 				product.gst_percent = gst_percent
 				product.price = price
 				db.session.commit()
+				db.session.add(ActivityLog(user_id=current_user.id, username=current_user.username, action='Edit Product', details=f'Product: {product.name}'))
+				db.session.commit()
 				flash('Product updated!', 'success')
 		else:
 			product = Product(name=name, hsn_code=hsn_code, gst_percent=gst_percent, price=price)
 			db.session.add(product)
 			db.session.commit()
+			db.session.add(ActivityLog(user_id=current_user.id, username=current_user.username, action='Add Product', details=f'Product: {product.name}'))
+			db.session.commit()
 			flash('Product added!', 'success')
 		return redirect(url_for('products.crud_products'))
 	# GET
+	search = request.args.get('search', '')
 	if edit_id:
 		form_product = Product.query.get(edit_id)
-	products = Product.query.order_by(Product.id).all()
-	return render_template('products/crud.html', products=products, form_product=form_product)
+	query = Product.query
+	if search:
+		query = query.filter(
+			Product.name.ilike(f'%{search}%') |
+			Product.hsn_code.ilike(f'%{search}%')
+		)
+	products = query.order_by(Product.id).all()
+	return render_template('products/crud.html', products=products, form_product=form_product, search=search)
